refactor: improve collapse implementation
