+++ src/components/collapse/index.ts
@@ -1,8 +1,9 @@
 import './collapse.less'
 import { attachPropertiesToComponent } from '../../utils/attach-properties-to-component'
-import { Collapse, CollapsePanel } from './collapse'
+import { Collapse } from './collapse'
+import { CollapsePanel } from './collapse-panel'
 
-export type { CollapseProps, CollapsePanelProps } from './collapse'
+export type { CollapseProps } from './collapse'
 
 export default attachPropertiesToComponent(Collapse, {
   Panel: CollapsePanel,
